Fixed some issues and implemented almost all passes into rendergraph
Changed up how the renderer differentiates between forward and deferred rendering

+++ Lamp/src/Lamp/Input/FileSystem.cpp
@@ -174,7 +174,7 @@ namespace Lamp
 
 		std::streampos end = stream.tellg();
 		stream.seekg(0, std::ios::beg);
-		uint32_t size = end - stream.tellg();
+		uint32_t size = (uint32_t)(end - stream.tellg());
 		LP_CORE_ASSERT(size != 0, "");
 
 		buffer.Allocate(size);
