+++ Sandbox/src/Sandbox3D/Windows/SandboxMeshImporter.cpp
@@ -377,7 +377,7 @@ namespace Sandbox3D
 			int matId = 0;
 			for (auto& mat : m_pModelToImport->GetMaterials())
 			{
-				for (size_t i = 0; i < shaders.size(); i++)
+				for (int i = 0; i < shaders.size(); i++)
 				{
 					if (mat.second->GetShader()->GetName() == shaders[i])
 					{
